Ajout des mofications

+++ src/co/kozao/kotask/controllers/ProjectAccessController.java
@@ -5,11 +5,11 @@ import java.util.List;
 
 import org.apache.log4j.Logger;
 
-import co.kozao.kotask.models.Project;
+import co.kozao.kotask.models.ProjectModel;
 import co.kozao.kotask.models.enums.ProjectStatus;
 //import co.kozao.kotask.models.Role;
 //import co.kozao.kotask.models.User;
-import co.kozao.kotask.services.implement.ProjectService;
+import co.kozao.kotask.services.ProjectService;
 import co.kozao.kotask.services.interfaces.ProjectServiceInterface;
 import co.kozao.kotask.utils.actionvalidation.ProjectActionValidationUtils;
 
@@ -17,9 +17,9 @@ public class ProjectAccessController {
 	private ProjectServiceInterface projectService = new ProjectService();
 	private static final Logger LOGGER = Logger.getLogger(ProjectAccessController.class);
 
-	public Project createProject(String name, String projectKey, String description, LocalDate startDate,
+	public ProjectModel createProject(String name, String projectKey, String description, LocalDate startDate,
 			LocalDate endDate, ProjectStatus status, int idProjectManager) {
-		Project project = new Project();
+		ProjectModel project = new ProjectModel();
 		project.setName(name);
 		project.setProjectKey(projectKey);
 		project.setDescription(description);
@@ -34,7 +34,7 @@ public class ProjectAccessController {
 
 			ProjectActionValidationUtils.validate(project);
 
-			Project createdProject = projectService.createProject(project);
+			ProjectModel createdProject = projectService.createProject(project);
 
 	        if (createdProject != null) {
 	            LOGGER.info("Projet créé avec succès : " + createdProject);
@@ -50,7 +50,7 @@ public class ProjectAccessController {
 		}
 	}
 
-	public boolean updateProject(Project project) {
+	public boolean updateProject(ProjectModel project) {
 		try {
 			ProjectActionValidationUtils.validate(project);
 			return projectService.updateProject(project);
@@ -62,15 +62,18 @@ public class ProjectAccessController {
 
 	public boolean deleteProject(int idProject) {
 		try {
-			projectService.deleteProject(idProject);
-			return true;
+			if(ProjectActionValidationUtils.validateidProject(idProject)) {
+				return false;
+			}
+			return 	projectService.deleteProject(idProject);
+			
 		} catch (Exception e) {
 			LOGGER.error("Erreur lors de la suppression du projet : " + e.getMessage(), e);
 			return false;
 		}
 	}
 
-	public Project getProjectById(int idProject) {
+	public ProjectModel getProjectById(int idProject) {
 		try {
 			return projectService.getProjectById(idProject);
 		} catch (Exception e) {
@@ -79,7 +82,7 @@ public class ProjectAccessController {
 		}
 	}
 
-	public List<Project> getAllProjects() {
+	public List<ProjectModel> getAllProjects() {
 		try {
 			return projectService.getAllProjects();
 		} catch (Exception e) {
